Get rid of package search, make naming consistent
The package search can at best sort for popularity
which is not necessarily the best option. For example,
`csv` is the most popular for Elixir, so that would
be recommended when using `package_search`, but if
the AI decides by itself (or use web search), then
it reaches for NimbleCSV.

Furthermore, for other ecosystems that don't provide
`search_package_docs`, without reading further pages,
the tool won't know how to use the package, so the
search by itself is not really relevant.